I FIXED THE TEXTUREHANDLER DEPENDECIES!!!!!!!!

+++ include/config.h
@@ -9,6 +9,6 @@ namespace Config {
 	const int screenHeight = 768;
 	const int volume = 50;
 	const int framerate = 60;
-	const double holdTimeThreshold = 100;
+	const float holdTimeThreshold = 100;
 	const SDL_WindowFlags screenType = SDL_WINDOW_SHOWN;
 }
